@@ -1,14 +1,14 @@
 package com.levimoreira.teammenagerapp.business.views
 
 
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProvider
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -55,7 +55,8 @@ class BusinessListFragment : DaggerFragment(), LifecycleOwner {
             Toast.makeText(this.context, business.title, Toast.LENGTH_SHORT).show()
         }
 
-        businessRecyclerView.layoutManager = LinearLayoutManager(this.context)
+        businessRecyclerView.layoutManager =
+            androidx.recyclerview.widget.LinearLayoutManager(this.context)
         businessRecyclerView.adapter = adapter
 
         addBusiness.setOnClickListener {
