@@ -1,6 +1,6 @@
 package com.levimoreira.teammenagerapp.organization.adapters
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 import com.levimoreira.teammenagerapp.R
@@ -8,7 +8,7 @@ import com.levimoreira.teammenagerapp.application.entities.Organization
 import com.levimoreira.teammenagerapp.inflate
 import kotlinx.android.synthetic.main.organization_item_view.view.*
 
-class OrganizationListAdapter(private val list: List<Organization>) : RecyclerView.Adapter<OrganizationListAdapter.OrganizationViewHolder>() {
+class OrganizationListAdapter(private val list: List<Organization>) : androidx.recyclerview.widget.RecyclerView.Adapter<OrganizationListAdapter.OrganizationViewHolder>() {
     lateinit var clickListener: (Organization) -> (Unit)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): OrganizationViewHolder = OrganizationViewHolder(parent.inflate(R.layout.organization_item_view))
@@ -21,7 +21,7 @@ class OrganizationListAdapter(private val list: List<Organization>) : RecyclerVi
     }
 
 
-    inner class OrganizationViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    inner class OrganizationViewHolder(itemView: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(itemView) {
         fun bind(item: Organization, clicklistener: (Organization) -> Unit) = with(itemView) {
             name.text = item.name
             phone.text = item.phone
