@@ -2,7 +2,7 @@ package com.levimoreira.teammenagerapp.task
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,7 +13,7 @@ import com.levimoreira.teammenagerapp.R
  * A simple [Fragment] subclass.
  *
  */
-class TasksFragment : Fragment() {
+class TasksFragment : androidx.fragment.app.Fragment() {
     val TAG = "TasksFragment"
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
