@@ -1,6 +1,6 @@
 package com.levimoreira.teammenagerapp.application.di
 
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProvider
 import dagger.Binds
 import dagger.Module
 
