@@ -1,9 +1,9 @@
 package com.levimoreira.teammenagerapp.application.entities
 
-import android.arch.persistence.room.ColumnInfo
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.ForeignKey
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
 import com.levimoreira.teammenagerapp.application.entities.Business.Companion.TABLE_NAME
 
 
