@@ -1,8 +1,9 @@
 package com.levimoreira.teammenagerapp.application.entities
 
-import android.arch.persistence.room.ColumnInfo
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.ColumnInfo
+import androidx.room.Embedded
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 import com.levimoreira.teammenagerapp.application.entities.Person.Companion.TABLE_NAME
 
 
@@ -10,7 +11,8 @@ import com.levimoreira.teammenagerapp.application.entities.Person.Companion.TABL
 data class Person(@PrimaryKey(autoGenerate = true) var id: Long?,
                   @ColumnInfo(name = COLUMN_NAME) var name: String,
                   @ColumnInfo(name = COLUMN_PHONE) var phone: String,
-                  @ColumnInfo(name = COLUMN_EMAIL) var email: String
+                  @ColumnInfo(name = COLUMN_EMAIL) var email: String,
+                  @Embedded var address: Address
 
 ) {
     companion object {
@@ -18,6 +20,5 @@ data class Person(@PrimaryKey(autoGenerate = true) var id: Long?,
         const val COLUMN_NAME = "name"
         const val COLUMN_PHONE = "phone"
         const val COLUMN_EMAIL = "email"
-
     }
 }
