@@ -1,11 +1,14 @@
 package com.levimoreira.teammenagerapp.business.data
 
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import com.levimoreira.teammenagerapp.application.entities.Business
+import com.levimoreira.teammenagerapp.application.entities.Business.Companion.COLUMN_PERSON
+import com.levimoreira.teammenagerapp.application.entities.BusinessPerson
 import com.levimoreira.teammenagerapp.application.entities.Organization
+import com.levimoreira.teammenagerapp.application.entities.Person
 import io.reactivex.Maybe
 import io.reactivex.Single
 
@@ -22,4 +25,14 @@ interface BusinessDao {
 
     @Query("DELETE FROM " + Business.TABLE_NAME + " WHERE id = :businessId")
     fun deleteById(businessId: Long): Int
+
+    @Query("SELECT * FROM " + Business.TABLE_NAME + " WHERE " + COLUMN_PERSON + " = :personId")
+    fun businessFromPerson(personId: Long): Single<List<Business>>
+
+    @Query("SELECT * FROM person INNER JOIN business_person ON person.id=business_person.personId WHERE business_person.businessId=:businessId")
+    fun getPersonForBusiness(businessId: Long): Single<List<Person>>
+
+    @Query("SELECT * FROM business INNER JOIN business_person ON business.id=business_person.businessId WHERE business_person.personId=:personId")
+    fun getBusinessForPerson(personId: Long): Single<List<Business>>
+
 }
