@@ -1,9 +1,9 @@
 package com.levimoreira.teammenagerapp.organization.data
 
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import com.levimoreira.teammenagerapp.application.entities.Organization
 import com.levimoreira.teammenagerapp.application.entities.Person
 import io.reactivex.Maybe
